requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
cffi==1.15.0
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
importlib-metadata==4.11.4
Markdown==3.3.7
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
zipp==3.8.0
